[lldb] Fix mangled name being overwritten on type with DW_AT_specification
The order of the attributes wasn't deterministic, and it just happened
that the DW_AT_linkage_name of the specifier type used to always come
last. Now the algorithm for parsing attributes has changed and made
deterministic, and the current algorithm will parse the attributes of
the specifier TAG first, and specified last.

Make sure to not overwrite the mangled name of a specified type.

rdar://158634690
(cherry picked from commit 3901846c9427aca7931713ced45cf7dc3d619940)