[lldb] Add extra verbose logging to flaky TestTsanSwiftAccessRace.py
We have seen another failure of TestTsanSwiftAccessRace.py, indicating that we recieved the stop reason eStopReasonSignal rather than the expected eStopReasonInstrumentation. This alone has not been enough to diagnose the root cause of the failure.

This PR adds in verbose logging, as well as capturing the value of the signal, in the hope of diagnosing the issue fully in the next occurrence.

(cherry picked from commit 173d5640ac019f8cf53556334a7236bc4aac5b5a)