Reland "[C++ BoundsSafety] Fix false positives when pointer argument is a function call (#11046)"  (#11358)
When the pointer argument passing to a counted_by parameter is a
function call returning a counted_by pointer, in order to check the
compatibility with respect to their counted_by attributes, the
compiler need to compare two expressions, which both have references
to parameters that need substitutions.

This commit is to generalize the compatibility checking to allow both
comparands to be accompanied by substitution maps.

rdar://155952016