[lldb] Fix ThreadPlanStepOut when returning from trampolines
When ThreadPlanStepOut decides to step out of a frame, it enqueues a
subplan to do so. However, when that plan is complete,
ThreadPlanStepOut::ShouldStop forget to set the `done` boolean properly.
This was exposed by a recent patch in LLVM:
https://github.com/llvm/llvm-project/pull/126838/files

(cherry picked from commit 9fb9567f859193d9ba43ce02c67a60d1642b874b)