Reland "Revert "[lldb] Fix OP_deref evaluation for large integer results (#159460)""
The original had an issue on "AArch-less" bots.
Fixed it with some ifdefs around the presence of the AArch ABI plugin.

Note for the cherry-pick: the test was removed as the related test file
in this branch is too old.

This reverts commit 1a4685d.

Cherry-picked from 40eb97623032ccf0a6a8f1adaea6e37014827f34.