Fix bug in cas-friendly-debug-info option.
Fix bug with emitting a DW_LNS_set_file at the beginning of a new
function's contribution to a line table, when the
-cas-friendly-debug-info option is passed. We were setting the current
FileNum to 0, however, if the LineEntry to be emitted also has the
FileNum be 0, we don't emit a DW_LNS_set_file at all, it is only emitted
if the current FileNum and the LineEntry to be emitted have different
FileNums.

To fix this, in cas-friendly-debug-info, we always set the current
FileNum to be -1.

 rdar://161930616

(cherry picked from commit 823563a6785866c4ded0a4b3fb44b24592f283f3)