[CAS] Cleanup chaining for UnifiedOnDiskCache
Previously, the chaining of KeyValueDB and OnDiskGraphDB is not
consistant. Some operations are implemented directly in the lowest
layer, some are in the UnifiedOnDiskCache layer, and some are in the
ActionCache/ObjectStore layer.

Now unifies all the chaining logics down into OnDiskGraphDB and
OnDiskKeyValueDB layer, with the exception of KeyValueDB chaining will
need the help of functions in UnifiedOnDiskCache layer.

This cleans up the interfaces for UnifiedOnDiskCache member functions
so it only contains database managment functions. Old functions like
`KVPut/Get` can be done directly via underlying database file with a
little bit of extra wrapper around it (see libCASPluginTest.dylib)
implementation for the simple wrapper needed.