[libclang][reproducers] Redirect writing output to the reproducer cache directory. (#11853)
The intention is to keep reproducers relocatable. Without redirection we are
trying to write output to "/some/path/existing/only/where/reproducer/generated.o"
and failing to do so. Instead we redirect the output to the reproducer cache
directory which is a relative path that is guaranteed to exist.

.pcm files are the output too but we don't redirect them because they
should be written and read from a specific place which is controlled by
`LookupOutput` callback.

(cherry picked from commit 87d0b80de80ede8ae4e30025ec58f5d26b715bc1)