[AArch64] Make StackTagging use alloca getAllocationSize rather than lifetime size
This was causing zero sized setTags to be emitted when compiling with Swift with memtag-stack sanitization; this caused an assert in the backend 'N->getOpcode() != ISD::EntryToken && "EntryToken in CSEMap!"'.

This patch does as the title says and replaces the use of the lifetime size argument with AllocaInst.getAllocationSize.

rdar://164400893