[VPlan] Handle min/max intrinsics in getSCEVExprForVPValue (NFCI)
Use m_Intrinsic to handle min/max intrinsics in getSCEVExprForVPValue.
This also extends Argument_match and IntrinsicID_match to VPInstruction
for completeness, and unifies the handling to avoid looking up functions
from the underlying IR instruction.

Tested via the VPlan-based cost-model, but same costs should be
computed.

As part of the extension, fix a bug in Argument_match that had an
incorrect offset for the operands of VPReplicateRecipe; the function is
the last argument.

(cherry picked from commit 7de080482c0cbbf01af05ddf648282591e5ed357)